Move variable comparison tests to file
